Generic assembler

This is the generic assembler interface and types; the idea is that the method are dispatched based on the basis type and the dynamic type.

RigorousInvariantMeasures.DiscretizedOperator — Type

This is an abstract type representing a discretized operator

RigorousInvariantMeasures.DiscretizedOperator — Method

DiscretizedOperator(B, D; ϵ = 10^(-14), max_iter = 100, T = Float64)

Constructor of a discretized operator, it infers if the operator is integral preserving or not from the basis B

Arguments: B Basis D Dynamic ϵ stopping condition for the Newton method max_iter maximum number of Newton method iterate T type of the elements stored in the matrix

RigorousInvariantMeasures.IntegralPreservingDiscretizedOperator — Type

This type represents an operator that preserves the value of the integral

RigorousInvariantMeasures.NonIntegralPreservingDiscretizedOperator — Type

An operator of the form Q = L + e*w (sparse plus rank-1); this 
is an operator that has been corrected to preserve the integral