These are some implemented methods for building dynamics

Abstract Dynamic Definition

RigorousInvariantMeasures.dfly_inf_der — Function

dfly inequality for maps with infinite derivatives.

The strategy to compute it follows a variant of Lemma 9.1 in the GMNP paper:

we find a "problematic set" I by taking a small interval of size radius(branch domain)/2^3 around each endpoint with infinite derivative;

we find l such that T >= l for each point in I

we compute the dfly coefficients as in the lemma.

we repeat the computation replacing 2^3 with 2^4, 2^5, ... 2^15 and take the best estimate among these.

RigorousInvariantMeasures.has_infinite_derivative_at_endpoints — Method

Returns a pair (left::Bool, right::Bool) that tells if a branch has infinite derivative at any of its endpoints

RigorousInvariantMeasures.DynamicDefinition.derivative — Method

derivative(n, f, x)

Nth derivative of a function (or a dynamic)

RigorousInvariantMeasures.DynamicDefinition.distortion — Method

distortion of a function (or a dynamic), i.e., |f′′ / f′^2|

RigorousInvariantMeasures.DynamicDefinition.endpoints — Method

endpoints(S::Dynamic)

Endpoints of the branches, in increasing order (returned as a vector of intervals)

RigorousInvariantMeasures.DynamicDefinition.expansivity — Function

Maximum of |1/T'|

RigorousInvariantMeasures.DynamicDefinition.max_distortion — Function

Maximum of distortion(D, x) = |T''| / (T')^2, over all branches

RigorousInvariantMeasures.DynamicDefinition.orientation — Function

orientation(D, k)

Orientation of branch k: 1. for increasing, -1. for decreasing

RigorousInvariantMeasures.DynamicDefinition.plottable — Method

Return a non-interval version of the map as a function. This can be used, for instance, for plot(plottable(D)).

RigorousInvariantMeasures.DynamicDefinition.preim — Function

preim(S::Dynamic, k, y, ϵ)

Computes the preim of y in branch k of a dynamic, with accuracy ϵ

Piecewise-Map Definition

RigorousInvariantMeasures.PwDynamicDefinition.PwMap — Type

Dynamic based on a piecewise monotonic map.

The map is defined as T(x) = Tsk if x ∈ [endpoints[k], endpoints[k+1]).

y_endpoints (kx2 matrix) contains the result of applying Ts to the endpoints of each interval. These can be filled in automatically from endpoints, but sometimes they are known to higher accuracy, for instance for x -> mod(3x, 1) we know that it is full-branch exactly. It is assumed that the map will send its domain hull(endpoints[begin],endpoints[end]) into itself.

the array branches is guaranteed to satisfy branches[i].X[end]==branches[i+1].X[begin]

RigorousInvariantMeasures.PwDynamicDefinition.PwMap — Method

Function call, and Taylor expansion, of a PwMap. Note that this ignores discontinuities; users are free to shoot themselves in the foot and call this on a non-smooth piecewise map. No better solutions for now.

RigorousInvariantMeasures.DynamicDefinition.branch — Method

function that evaluates the k-th branch of a dynamic on a point x (assuming it's in its domain, otherwise ∅)

RigorousInvariantMeasures.DynamicDefinition.distortion — Method

Return the distortion function of a branch, i.e., |f′′ / f′^2|

RigorousInvariantMeasures.DynamicDefinition.expansivity — Function

expansivity(D::PwMap; tol=1e-3)

Compute a rigorous bound for the expansivity of a PwMap

RigorousInvariantMeasures.DynamicDefinition.max_distortion — Function

max_distortion(D::PwMap; tol=1e-3)

Compute a rigorous bound for the distortion of a PwMap

Example

julia> using RigorousInvariantMeasures;

julia> D0 = mod1_dynamic(x->2*x+0.5*x*(1-x), full_branch = true)
Piecewise-defined dynamic with 2 branches

julia> max_distortion(D0)
[0.444268, 0.444445]

RigorousInvariantMeasures.PwDynamicDefinition.bound_branch_distortion — Function

bound_branch_distortion(br::Branch; tol = 0.01)

Compute a rigorous bound for the distortion of a branch on an interval I, defaults to the domain of the branch

RigorousInvariantMeasures.PwDynamicDefinition.branch_expansivity — Function

branch_expansivity(br::Branch; tol = 0.01)

Compute a rigorous bound for the expansivity of a branch

RigorousInvariantMeasures.PwDynamicDefinition.restrict — Method

Intersect an Interval or TaylorSeries with I