Generic estimate

This is the generic estimator interface and types; the idea is that the method are dispatched based on the basis type and the dynamic type.

RigorousInvariantMeasures.CoarseGridQuantities — Type

Struct that encapsulates the additional quantities needed on the coarse basis for a two-grid estimate, or on the (only) basis for a one-grid estimate. It is meant as an intermediate quantity that can be saved on the disk to avoid recomputing Q all the times.

RigorousInvariantMeasures.FineGridQuantities — Type

Struct that encapsulates all the quantities computed from the fine basis that are needed in the two-grid estimate. It is meant as an intermediate quantity that can be saved on the disk to avoid recomputing Q all the times

RigorousInvariantMeasures.boundnorm — Method

This function returns the bound on the weak norm of the discretized operator

RigorousInvariantMeasures.compute_coarse_grid_quantities — Method

Compute FineGridQuantities and CoarseGridQuantities, given a function f(n) that computes B, D, Q = f(n)

RigorousInvariantMeasures.compute_fine_grid_quantities — Method

Compute FineGridQuantities, given a function f(n) that computes B, D, Q = f(n)

RigorousInvariantMeasures.distance_from_invariant — Method

Bounds rigorously the distance of w from the fixed point of Q (normalized with integral = 1), using a vector of bounds norms[k] ≥ ||Qh^k|{U_h^0}||. If ε₁ and normQ are given, then Q can be omitted

RigorousInvariantMeasures.finepowernormbounds — Method

Uses power norm bounds already computed for a coarse operator to estimate the same norms for a finer operator

RigorousInvariantMeasures.invariant_vector — Method

Return a numerical approximation to the (hopefully unique) invariant vector of the dynamic with discretized operator Q.

The vector is normalized so that integral_covector(B)*w ≈ 1

RigorousInvariantMeasures.one_grid_estimate — Method

Compute a one-grid error estimate.

The first return argument is the error, the second is the time breakdown according to ["dfly", "assembling", "eigen", "norms", "error"]. (The sum of that vector is the total time taken)

RigorousInvariantMeasures.powernormbounds — Method

Uses different strategies to compute power norm bounds.

If specified, m norms of powers are estimated computationally, and then m_extend norms are obtained with a cheaper refinement process. Otherwise these numbers are selected automatically.

A vector of length mextend is returned, such that norms[k] ≥ ||Qh^k|{Uh^0}||

RigorousInvariantMeasures.powernormbounds — Method

Computes bounds for norms of powers, taking (optionally) minimum values for the number of norms to compute

RigorousInvariantMeasures.residualbound — Method

Return an upper bound to Q_h*w - w in the given norm

RigorousInvariantMeasures.two_grid_estimate — Method

Compute a two-grid error estimate.

The first return argument is the error, the second is the time breakdown according to ["dfly", "coarse", "assembling", "eigen", "norms", "error"]. (The sum of that vector is the total time taken)

DFLY

RigorousInvariantMeasures.SymbL — Type

Represent a symbolic transfer operator
`(L_n f)(x) = \sum_{y\in T^{-1}(x)}\frac{f(y)}{|T'(y)|^n}`

RigorousInvariantMeasures.Diff — Method

Compute the derivative of a symbolic transfer operator  
`P` with respect to differential `∂`
and with derivative of `1/T′ ` given by the function 
`Dist`.
Refer to Eq. 3.2 in Butterley-Kiamari-Liverani Locating Ruelle Pollicot Resonances

RigorousInvariantMeasures.Diff — Method

Given the summands of a sum of Symbolic Transfer Operator
given in a vector, computes a vector containing 
the sum of the derivatives

RigorousInvariantMeasures.dfly — Method

(A, B) = dfly(strongnorm, auxnorm, dynamic)

Constants (A, B) such that ||Lf||s ≦ A||f||s + B||f||_aux