Library

Contractors

These are interval contractors implemented here; in the future they may be substituted by established libraries, as IntervalRootFinding.jl

RigorousInvariantMeasures.Contractors.nthpreimage! — Method

Newer version of the 'shooting method' to compute the kth preimage of a point (or interval y) fs contains k functions, X contains their domains. This computes a solution of fk(f{k-1}( ... f1(x) ... )) = y. Overwrites X with [x f1(x) f2(f1(x)) ... f{k-1}(...)], so the true solution is X[1].

Tries to avoid allocations and stuff.

RigorousInvariantMeasures.Contractors.root — Method

Compute a single root with (possibly multivariate) interval Newton

x must be an Interval (univariate) or IntervalBox (multivariate)

Stops when the interval reaches a fixed point, when the diameter is smaller than ε, or when max_iter iterations are reached (with an error)

RigorousInvariantMeasures.Contractors.unique_increasing — Method

unique_increasing(a, b)

Given intervals a, b, returns true if a < b, false if b < a, and raises an error if it is not uniquely determined.

RigorousInvariantMeasures.Contractors.unique_sign — Method

unique_sign(x)

Sign of an interval, but throws an error if it is not unique. Used by various functions to compute orientations

Preimages

Optimized methods to compute preimages of $1$ dimensional maps.

RigorousInvariantMeasures.ComposedDynamic — Type

Composed map D1 ∘ D2 ∘ D3. We store with [D1, D2, D3] in this order.

We overwrite ∘ in base, so one can simply write D1 ∘ D2 or ∘(D1, D2, D3) to construct them.

RigorousInvariantMeasures.DynamicDefinition.domain — Method

Utility function to return the domain of a dynamic

RigorousInvariantMeasures.first_overlapping — Method

first_overlapping(y, a)

Smallest possible i such that a is in the semi-open interval [y[i], y[i+1]).

This should work properly even if a, y are intervals; in this case it returns the smallest possible value of i over all possible "assignments" of a, y inside those intervals. Assumes y is sorted, i.e., map(y, x->Interval(x).lo) and map(y, x->Interval(x).hi) are sorted.

RigorousInvariantMeasures.last_overlapping — Method

last_overlapping(y, a)

Largest possible j such that a-ε is in the semi-open interval [y[j], y[j+1]).

This should work properly even if a, y are intervals; in this case it returns the largest possible value of i over all possible "assignments" of a, y inside those intervals. Assumes y is sorted, i.e., map(y, x->Interval(x).lo) and map(y, x->Interval(x).hi) are sorted.

RigorousInvariantMeasures.preimages — Function

preimages(y, br::Branch, ylabel = 1:length(y), ϵ = 0.0)

Construct preimages of an increasing array y under a monotonic branch defined on X = (a, b), propagating additional labels ylabel

The sequence y subdivides the y-axis into semi-open intervals [y[l], y[l+1]); each of them is identified by the label ylabel[l]. We construct an increasing sequence x that splits X (in the x-axis) into semi-open intervals, each of them with f([x[k], x[k+1]) ⊂ [y[l], y[l+1]) for a certain l. We set xlabel[k] = ylabel[l], and return the pair (x, xlabel).

It is assumed that it cannot happen that f(x) < y[1].

In the simplest case where D is full-branch, the points in x are preimages of the points in y, but in the general case they can also include D.endpoints: in general, there may be a certain number of points in y that have no preimage at the beginning and the end of the sequence, because they fall out of the range R = [f(a), f(b)]. In the worst case, no point has a preimage, because y[i] < R < y[i+1] for some i (or vice versa with orientations), and in this case we just return the 1-element vectors x = [branch.X[1]] and xlabel = [i].

x[begin] always coincides with branch.X[1], while branch.X[2] is "the point after x[end]", and is not stored explicitly in x, for easier composing. In this way x and xlabel have the same length.

This function fills the array by using a bisection strategy to save computations: if y ∈ [a,b], then f⁻¹(y) ∈ f⁻¹(a),f⁻¹(b). So we can fill v by filling in first entries v[k+1] with higher dyadic valuation of k.

For a dynamic with multiple branches, preimages(y, D) is simply the concatenation of x, xlabel for b in all branches. These values still form an increasing sequence that splits X into intervals, each of which is mapped into a different semi-open interval [y[k], y[k+1]).

RigorousInvariantMeasures.preimages — Function

preimages(y, D::Dynamic, ylabel = 1:length(y), ϵ = 0.0; progress = true)

Construct preimages of an increasing array y under a dynamic, propagating additional labels `ylabel`

RigorousInvariantMeasures.preimages_and_derivatives — Function

preimages_and_derivatives(y, br::Branch, ylabel = 1:length(y), ϵ = 0.0)

Compute preimages of D and the derivatives f'(x) in each point.

Returns: x, xlabel, x′

Assumes that the dynamic is full-branch, because otherwise things may compose the wrong way. This is not restrictive because we'll need it only for the Hat assembler (at the moment)

We combine them in a single function because there are avenues to optimize by recycling some computations (not completely exploited for now)

Index

RigorousInvariantMeasures.Dual

RigorousInvariantMeasures.Hat

RigorousInvariantMeasures.Hat

RigorousInvariantMeasures.HatFunction

RigorousInvariantMeasures.HatFunctionOnTorus

RigorousInvariantMeasures.IntervalOnTorus

RigorousInvariantMeasures.Ulam

RigorousInvariantMeasures.Ulam

RigorousInvariantMeasures._eval_T

RigorousInvariantMeasures.eval_Clenshaw_BackwardFirst

RigorousInvariantMeasures.get_norm

RigorousInvariantMeasures.relative_measure

RigorousInvariantMeasures.ComposedDynamic

RigorousInvariantMeasures.first_overlapping

RigorousInvariantMeasures.last_overlapping

RigorousInvariantMeasures.preimages

RigorousInvariantMeasures.preimages

RigorousInvariantMeasures.preimages_and_derivatives