Library

Contractors

These are interval contractors implemented here; in the future they may be substituted by established libraries, as IntervalRootFinding.jl

RigorousInvariantMeasures.Contractors.preimage_monotonic — Method

Compute the guaranteed interval preimage f⁻¹(y) of a point y ∈ R according to the function f: [a,b] → R. This preimage is guaranteed to lie inside the real interval x = hull(a,b).

f must be strictly monotonic on [a,b], at least in the sense of MonotonicBranch functions (allowing for uncertainty on the endpoints). f must be differentiable; zero and infinite derivative are allowed.

y1, y2 must be guaranteed to be equal to f(a), f(b) or "more outside", so that y ∉ hull(y1,f(z)) ⟹ f⁻¹(y) ∉ hull(a, z).

Stops when the interval reaches a fixed point, when the diameter is smaller than ε, or when max_iter iterations are reached (with an error)

RigorousInvariantMeasures.Contractors.unique_increasing — Method

unique_increasing(a, b)

Given intervals a, b, returns true if a < b, false if b < a, and raises an error if it is not uniquely determined.

Preimages

Optimized methods to compute preimages of $1$ dimensional maps.

RigorousInvariantMeasures.ComposedDynamic — Type

Composed map D1 ∘ D2 ∘ D3. We store with [D1, D2, D3] in this order.

We overwrite ∘ in base, so one can simply write D1 ∘ D2 or ∘(D1, D2, D3) to construct them.

RigorousInvariantMeasures.DynamicDefinition.domain — Method

Utility function to return the domain of a dynamic

RigorousInvariantMeasures.first_overlapping — Method

first_overlapping(y, a)

Smallest possible i such that a is in the semi-open interval [y[i], y[i+1]).

This should work properly even if a, y are intervals; in this case it returns the smallest possible value of i over all possible "assignments" of a, y inside those intervals. Assumes y is sorted, i.e., map(y, x->Interval(x).lo) and map(y, x->Interval(x).hi) are sorted.

RigorousInvariantMeasures.last_overlapping — Method

last_overlapping(y, a)

Largest possible j such that a-ε is in the semi-open interval [y[j], y[j+1]).

This should work properly even if a, y are intervals; in this case it returns the largest possible value of i over all possible "assignments" of a, y inside those intervals. Assumes y is sorted, i.e., map(y, x->Interval(x).lo) and map(y, x->Interval(x).hi) are sorted.

RigorousInvariantMeasures.preimage — Method

Compute the preimage f⁻¹(y), knowing that it lies inside search_interval.

RigorousInvariantMeasures.preimages — Function

preimages(y, br::MonotonicBranch, ylabel = 1:length(y); ϵ, max_iter)

Construct preimages of an increasing array y under a monotonic branch defined on X = (a, b), propagating additional labels ylabel

Construct preimages of an increasing array y under a monotonic branch br defined on X = (a, b), propagating additional labels ylabel

It is assumed that it cannot happen that $f(x) < y[1]$.

Extended help

The sequence y subdivides the y-axis into semi-open intervals [y[l], y[l+1]); each of them is identified by the label ylabel[l]. We construct an increasing sequence x that splits X (in the x-axis) into semi-open intervals, each of them with f([x[k], x[k+1]) ⊂ [y[l], y[l+1]) for a certain l. We set xlabel[k] = ylabel[l], and return the pair (x, xlabel).

It is assumed that it cannot happen that f(x) < y[1].

In the simplest case where D is full-branch, the points in x are preimages of the points in y, but in the general case they can also include D.endpoints: in general, there may be a certain number of points in y that have no preimage at the beginning and the end of the sequence, because they fall out of the range R = [f(a), f(b)]. In the worst case, no point has a preimage, because y[i] < R < y[i+1] for some i (or vice versa with orientations), and in this case we just return the 1-element vectors x = [branch.X[1]] and xlabel = [i].

x[begin] always coincides with branch.X[1], while branch.X[2] is "the point after x[end]", and is not stored explicitly in x, for easier composing. In this way x and xlabel have the same length.

This function fills the array by using a bisection strategy to save computations: if y ∈ [a,b], then f⁻¹(y) ∈ f⁻¹(a),f⁻¹(b). So we can fill v by filling in first entries v[k+1] with higher dyadic valuation of k.

For a dynamic with multiple branches, preimages(y, D) is simply the concatenation of x, xlabel for b in all branches. These values still form an increasing sequence that splits X into intervals, each of which is mapped into a different semi-open interval [y[k], y[k+1]).

Example

julia> using RigorousInvariantMeasures

julia> D = mod1_dynamic(x->2x)
Piecewise-defined dynamic with 2 branches

julia> RigorousInvariantMeasures.preimages(0:0.1:1, D.branches[1]; ϵ = 10^(-15), max_iter = 100)
(Interval{Float64}[[0, 0], [0.05, 0.0500001], [0.1, 0.100001], [0.149999, 0.15], [0.2, 0.200001], [0.25, 0.25], [0.299999, 0.3], [0.349999, 0.35], [0.4, 0.400001], [0.45, 0.450001]], [1, 2, 3, 4, 5, 6, 7, 8, 9, 10])

RigorousInvariantMeasures.preimages — Function

preimages(y, D::Dynamic, ylabel = 1:length(y), ϵ = 0.0; progress = true)

Construct preimages of an increasing array y under a dynamic, propagating additional labels `ylabel`

RigorousInvariantMeasures.preimages_and_derivatives — Function

preimages_and_derivatives(y, br::MonotonicBranch, ylabel = 1:length(y), ϵ = 0.0)

Compute preimages of D and the derivatives f'(x) in each point.

Returns: x, xlabel, x′

Assumes that the dynamic is full-branch, because otherwise things may compose the wrong way. This is not restrictive because we'll need it only for the Hat assembler (at the moment)

We combine them in a single function because there are avenues to optimize by recycling some computations (not completely exploited for now)

RigorousInvariantMeasures.range_estimate_monotone — Method

Utility function that estimates the range of a monotone function

Index

RigorousInvariantMeasures.Dual

RigorousInvariantMeasures.Hat

RigorousInvariantMeasures.Hat

RigorousInvariantMeasures.HatFunction

RigorousInvariantMeasures.HatFunctionOnTorus

RigorousInvariantMeasures.IntervalOnTorus

RigorousInvariantMeasures.Ulam

RigorousInvariantMeasures.Ulam

RigorousInvariantMeasures._eval_T

RigorousInvariantMeasures.eval_Clenshaw_BackwardFirst

RigorousInvariantMeasures.get_norm

RigorousInvariantMeasures.relative_measure

RigorousInvariantMeasures.ComposedDynamic

RigorousInvariantMeasures.first_overlapping

RigorousInvariantMeasures.last_overlapping

RigorousInvariantMeasures.preimage

RigorousInvariantMeasures.preimages

RigorousInvariantMeasures.preimages

RigorousInvariantMeasures.preimages_and_derivatives

RigorousInvariantMeasures.range_estimate_monotone

RigorousInvariantMeasures.composedPwMap

RigorousInvariantMeasures.dfly_inf_der

RigorousInvariantMeasures.has_infinite_derivative_at_endpoints

RigorousInvariantMeasures.mod1_dynamic