Library

Module

RigorousInvariantMeasures.CoarseGridQuantities — Type

Struct that encapsulates the additional quantities needed on the coarse basis for a two-grid estimate, or on the (only) basis for a one-grid estimate. It is meant as an intermediate quantity that can be saved on the disk to avoid recomputing Q all the times.

RigorousInvariantMeasures.ComposedDynamic — Type

Composed map D1 ∘ D2 ∘ D3. We store with [D1, D2, D3] in this order.

We overwrite ∘ in base, so one can simply write D1 ∘ D2 or ∘(D1, D2, D3) to construct them.

RigorousInvariantMeasures.Dual — Type

Replacement of DualComposedWithDynamic.

RigorousInvariantMeasures.FineGridQuantities — Type

Struct that encapsulates all the quantities computed from the fine basis that are needed in the two-grid estimate. It is meant as an intermediate quantity that can be saved on the disk to avoid recomputing Q all the times

RigorousInvariantMeasures.Fourier1D — Method

Fourier(N) builds a Fourier basis truncating at frequence N by default it uses a FFT of size 2^(log2(ceil(N))+2)

Fourier(N, FFTNx) builds a Fourier basis with FFT size FFTNx

The base size is 2*N+1 (0, ..., N, -N, ..., -1)

RigorousInvariantMeasures.Hat — Type

This type represents a Hat basis on $S^1$. It contains a vector with the midpoints of the hats.

RigorousInvariantMeasures.Hat — Method

This constructs a Hat basis on $S^1$ on equispaced points

RigorousInvariantMeasures.HatFunction — Type

Hat function (on the reals)

This is a piecewise linear function such that: f(x) = 0 if x <= lo f(mi) = 1 f(hi)

RigorousInvariantMeasures.HatFunctionOnTorus — Type

Hat function (on the torus)

This is a piecewise linear function such that: f(x) = 0 if x <= lo f(mi) = 1 f(x) = 0 if x >= ho

RigorousInvariantMeasures.IntervalOnTorus — Type

A separate type for intervals on the torus (mod 1) to "remind" us of the quotient

The interval is normalized in the constructor: the caller may assume that

0 <= i.lo < 1

i.hi < i.lo + 1 OR i==Interval(0,1)

RigorousInvariantMeasures.NonIntegralPreservingDiscretizedOperator — Type

An operator of the form Q = L + e*w (sparse plus rank-1).

RigorousInvariantMeasures.NormCacher — Type

Types to compute norms iteratively by "adding a column at a time".

RigorousInvariantMeasures.NormCacher — Method

Create a new NormCacher to compute the normbound of the empty matrix with n rows

RigorousInvariantMeasures.Ulam — Type

Equispaced Ulam basis on [0,1] of size n

Base.getindex — Method

makes so that B[j] returns a HatFunctionOnTorus with the j-th basis element

Base.getindex — Method

makes so that B[j] returns a HatFunctionOnTorus with the j-th basis element

Base.getindex — Method

makes so that B[j] returns a HatFunction with the j-th basis element

Base.iterate — Method

Given a preimage of an interval I_i, this iterator returns its relative intersection with all the elements of the Ulam basis that have nonzero intersection with it

Base.iterate — Method

Given a preimage y of a point x, this iterator returns \phi_j(y)/T'(y)

Base.iterate — Method

Given a preimage y of a point x, this iterator returns \phi_j(y)/T'(y)

Base.length — Method

Return the size of the Chebyshev basis

Base.length — Method

Return the size of the Hat basis

Base.length — Method

Return the size of the Hat basis

RecipesBase.apply_recipe — Function

Displays error on a function in the Ulam basis

The w argument is unused, but kept for compatibility with other functions for different bases

RecipesBase.apply_recipe — Method

Plots a function in the Hat basis

RecipesBase.apply_recipe — Method

Displays error on a function in the Hat basis

RecipesBase.apply_recipe — Method

Plots a function in the Hat basis

RecipesBase.apply_recipe — Method

Displays error on a function in the Hat basis

RecipesBase.apply_recipe — Method

Plots a function in the Ulam basis

RigorousInvariantMeasures.BasisDefinition.nonzero_on — Method

Return the range of indices of the elements of the basis whose support intersects with the given dual element (i.e., a pair (y, absT')). The range may end with length(B)+1; this must be interpreted "mod length(B)": it means that it intersects with the hat function peaked in 0 as well (think for instance y = 0.9999).

RigorousInvariantMeasures.BasisDefinition.nonzero_on — Method

Return the range of indices of the elements of the basis whose support intersects with the given dual element (i.e., a pair (y, absT')). The range may end with length(B)+1; this must be interpreted "mod length(B)": it means that it intersects with the hat function peaked in 0 as well (think for instance y = 0.9999).

RigorousInvariantMeasures.BasisDefinition.nonzero_on — Method

Returns the indices of the elements of the Ulam basis that intersect with the interval y We do not assume an order of a and b; this should not matter unless the preimages are computed with very low accuracy We assume, though, that y comes from the (possibly inexact) numerical approximation of an interval in [0,1], i.e., we restrict to y ∩ [0,1]

RigorousInvariantMeasures.BasisDefinition.normbound — Method

Rigorous upper bound on a vector norm. Note that Linf, L1 are the "analyst's" norms

RigorousInvariantMeasures.BasisDefinition.opnormbound — Method

This function bounds the L2 operator norm of M (complex floating point matrix)

RigorousInvariantMeasures.BasisDefinition.opnormbound — Method

Certified upper bound to ||A|| (of specified NormKind)

RigorousInvariantMeasures.BasisDefinition.opnormbound — Method

These functions compute a rigorous upper bound for the 2-norm of a vector; we have a specialized version for complex numbers to avoid taking the sqrt root and squaring again

RigorousInvariantMeasures.DynamicDefinition.domain — Method

Utility function to return the domain of a dynamic

RigorousInvariantMeasures._eval_T — Method

Eval the Chebyshev polynomial up to degree n on an array of points in [-1, 1].

Not satisfactory, the intervals explode

RigorousInvariantMeasures.abs_or_mag — Method

'Absolute value' definition that returns mag(I) for an interval and abs(x) for a real

RigorousInvariantMeasures.add_column! — Method

Update a NormCacher to add one column to the matrix it is computing a norm of. This column may be affected by an error ε (in the same norm).

RigorousInvariantMeasures.assemble — Function

Assembles the truncated Fourier matrix for a dynamic D, with frequencies -Nx≤ i≤ Nx

Input: B:: Fourier1D basis D: Dynamic T: Floating point type, supports Float64, BigFloat

Output:
M: matrix of intervals that rigorously contain M

RigorousInvariantMeasures.boundnorm — Method

This function returns the bound on the weak norm of the discretized operator

RigorousInvariantMeasures.compute_coarse_grid_quantities — Method

Compute FineGridQuantities and CoarseGridQuantities, given a function f(n) that computes B, D, Q = f(n)

RigorousInvariantMeasures.compute_fine_grid_quantities — Method

Compute FineGridQuantities, given a function f(n) that computes B, D, Q = f(n)

RigorousInvariantMeasures.dfly — Method

(A, B) = dfly(strongnorm, auxnorm, dynamic)

Constants (A, B) such that ||Lf||s ≦ A||f||s + B||f||_aux

RigorousInvariantMeasures.distance_from_invariant — Method

Bounds rigorously the distance of w from the fixed point of Q (normalized with integral = 1), using a vector of bounds norms[k] ≥ ||Qh^k|{U_h^0}||. If ε₁ and normQ are given, then Q can be omitted

RigorousInvariantMeasures.distance_from_invariant_noise — Method

Bounds rigorously the distance of w from the fixed point of Q (normalized with integral = 1), using a vector of bounds norms[k] ≥ ||Qh^k|{U_h^0}||.

RigorousInvariantMeasures.eig_costants_small_matrix — Method

eig_costants_small_matrix(A)

Return the dominant eigenvalue and associated left eigenvector of the small matrix

RigorousInvariantMeasures.eval_Clenshaw_BackwardFirst — Method

Eval a polynomial in Chebyshev basis, ClenshawBackward, using ball arithmetic Following Viviane Ledoux, Guillaume Moroz "Evaluation of Chebyshev polynomials on intervals andapplication to root finding"

RigorousInvariantMeasures.faster_certify — Method

This function uses the classical bound, cited in

RESIDUAL BOUNDS ON APPROXIMATE EIGENSYSTEMS OF NONNORMAL MATRICES* W. KAHAN, B. N. PARLETT AND E. JIANGt

Theorem 1

RigorousInvariantMeasures.finepowernormbounds — Method

Uses power norm bounds already computed for a coarse operator to estimate the same norms for a finer operator

RigorousInvariantMeasures.finepowernormboundsnoise — Method

Uses power norm bounds already computed for a coarse operator to estimate the same norms for a finer operator

RigorousInvariantMeasures.first_overlapping — Method

Smallest possible i such that a is in the semi-open interval [y[i], y[i+1]).

This should work properly even if a, y are intervals; in this case it returns the smallest possible value of i over all possible "assignments" of a, y inside those intervals. Assumes y is sorted, i.e., map(y, x->Interval(x).lo) and map(y, x->Interval(x).hi) are sorted.

RigorousInvariantMeasures.gamma — Method

γₙ constants for floating point error estimation, as in [Higham, Accuracy and Stability of Numerical Algorithms]

RigorousInvariantMeasures.get_norm — Method

Return the norm of the matrix the NormCacher is working on.

RigorousInvariantMeasures.get_norm — Method

Return the norm of the matrix the NormCacher is working on.

RigorousInvariantMeasures.infinite_sum_norms — Method

Estimate ||I + Q + Q^2 + … || (infinite sum) using a list of computed norm bounds norms[k] ≥ ||Q^k||.

RigorousInvariantMeasures.invariant_vector — Method

Return a numerical approximation to the (hopefully unique) invariant vector of the dynamic with discretized operator Q.

The vector is normalized so that integral_covector(B)*w ≈ 1

RigorousInvariantMeasures.invariant_vector_noise — Method

Return a numerical approximation to the (hopefully unique) invariant vector of the dynamic with discretized operator Q.

The vector is normalized so that integral_covector(B)*w ≈ 1

RigorousInvariantMeasures.last_overlapping — Method

Largest possible j such that a-ε is in the semi-open interval [y[j], y[j+1]).

This should work properly even if a, y are intervals; in this case it returns the largest possible value of i over all possible "assignments" of a, y inside those intervals. Assumes y is sorted, i.e., map(y, x->Interval(x).lo) and map(y, x->Interval(x).hi) are sorted.

RigorousInvariantMeasures.max_nonzeros_per_row — Method

Returns the maximum number of (structural) nonzeros in a row of A

RigorousInvariantMeasures.norm_2_estimator_square_nonrig — Method

This function computes a candidate for the top singular value of M|{V0} square, i.e., it returns λ^2 and the associated singular vector

RigorousInvariantMeasures.norms_of_powers — Method

Estimates the norms ||Q||, ||Q^2||, ... ||Q^m|| on U^0.

U is the matrix [ones(1,n-1); -I_(n-1,n-1)]. It is currently assumed that f*U==0 (i.e., all elements of f are equal).

f must be an interval vector.

The following constants may be specified as keyword arguments:

normQ, normE, normv0, normEF, normIEF, normN

otherwise they are computed (which may be slower).

e and f must be specified in case isintegralpreserving==false In case isintegralpreserving is true, they may be specified but they are then ignored.

Implementation note: currently we perform this computation one column at a time, to be able to scale (slowly) to cases with large size; for moderate sizes, it would indeed be better to do the computation all columns at the same time, in BLAS level 3.

RigorousInvariantMeasures.norms_of_powers_abstract_noise — Method

Arrays of bounds to ||Q^k||{w → s} = sup{||f||w=1} ||Q^k f||s and to ||Q^k||_{w} coming theoretically from iterated DFLY inequalities (the "small matrix method").

Returns two arrays (strongs, norms) of length m: strongs[k] bounds ||Q^k f||_s, norms[k] bounds ||Q^k f||)

RigorousInvariantMeasures.norms_of_powers_dfly — Method

Arrays of bounds to ||Q^k||{w → s} = sup{||f||w=1} ||Q^k f||s and to ||Q^k||_{w} coming theoretically from iterated DFLY inequalities (the "small matrix method").

Returns two arrays (strongs, norms) of length m: strongs[k] bounds ||Q^k f||_s, norms[k] bounds ||Q^k f||)

RigorousInvariantMeasures.norms_of_powers_from_coarser_grid — Method

Estimate norms of powers from those on a coarser grid (see paper for details)

RigorousInvariantMeasures.norms_of_powers_from_coarser_grid_noise — Method

Estimate norms of powers from those on a coarser grid (see paper for details) TODO: Check the final formula!!!

RigorousInvariantMeasures.norms_of_powers_from_coarser_grid_noise — Method

Estimate norms of powers from those on a coarser grid (see paper for details) TODO: Check if it works for other basis types

RigorousInvariantMeasures.norms_of_powers_noise — Method

Estimates the norms ||Q||, ||Q^2||, ... ||Q^m|| on U^0.

U is the matrix [ones(1,n-1); -I_(n-1,n-1)]. It is currently assumed that f*U==0 (i.e., all elements of f are equal).

f must be an interval vector.

The following constants may be specified as keyword arguments:

normQ, normE, normv0, normEF, normIEF, normN

otherwise they are computed (which may be slower).

e and f must be specified in case isintegralpreserving==false In case isintegralpreserving is true, they may be specified but they are then ignored.

RigorousInvariantMeasures.norms_of_powers_trivial — Method

Array of "trivial" bounds for the powers of a DiscretizedOperator (on the whole space) coming from from ||Q^k|| ≤ ||Q||^k

RigorousInvariantMeasures.norms_of_powers_trivial_noise — Method

Array of "trivial" bounds for the powers of a DiscretizedOperator (on the whole space) coming from from ||Q^k|| ≤ ||Q||^k

RigorousInvariantMeasures.norms_of_powers_trivial_noise — Method

Array of "trivial" bounds for the powers of a DiscretizedOperator (on the whole space) coming from from ||Q^k|| ≤ ||Q||^k

RigorousInvariantMeasures.one_grid_estimate — Method

Compute a one-grid error estimate.

The first return argument is the error, the second is the time breakdown according to ["dfly", "assembling", "eigen", "norms", "error"]. (The sum of that vector is the total time taken)

RigorousInvariantMeasures.powernormbounds — Method

Uses different strategies to compute power norm bounds.

If specified, m norms of powers are estimated computationally, and then m_extend norms are obtained with a cheaper refinement process. Otherwise these numbers are selected automatically.

A vector of length mextend is returned, such that norms[k] ≥ ||Qh^k|{Uh^0}||

RigorousInvariantMeasures.powernormbounds — Method

Computes bounds for norms of powers, taking (optionally) minimum values for the number of norms to compute

RigorousInvariantMeasures.preimages — Function

Construct preimages of an increasing array y under a monotonic branch defined on X = (a, b), propagating additional labels ylabel

The sequence y subdivides the y-axis into semi-open intervals [y[l], y[l+1]); each of them is identified by the label ylabel[l]. We construct an increasing sequence x that splits X (in the x-axis) into semi-open intervals, each of them with f([x[k], x[k+1]) ⊂ [y[l], y[l+1]) for a certain l. We set xlabel[k] = ylabel[l], and return the pair (x, xlabel).

It is assumed that it cannot happen that f(x) < y[1].

In the simplest case where D is full-branch, the points in x are preimages of the points in y, but in the general case they can also include D.endpoints: in general, there may be a certain number of points in y that have no preimage at the beginning and the end of the sequence, because they fall out of the range R = [f(a), f(b)]. In the worst case, no point has a preimage, because y[i] < R < y[i+1] for some i (or vice versa with orientations), and in this case we just return the 1-element vectors x = [branch.X[1]] and xlabel = [i].

x[begin] always coincides with branch.X[1], while branch.X[2] is "the point after x[end]", and is not stored explicitly in x, for easier composing. In this way x and xlabel have the same length.

This function fills the array by using a bisection strategy to save computations: if y ∈ [a,b], then f⁻¹(y) ∈ f⁻¹(a),f⁻¹(b). So we can fill v by filling in first entries v[k+1] with higher dyadic valuation of k.

For a dynamic with multiple branches, preimages(y, D) is simply the concatenation of x, xlabel for b in all branches. These values still form an increasing sequence that splits X into intervals, each of which is mapped into a different semi-open interval [y[k], y[k+1]).

RigorousInvariantMeasures.preimages_and_derivatives — Function

Compute preimages of D and the derivatives f'(x) in each point.

Returns: x, xlabel, x′

Assumes that the dynamic is full-branch, because otherwise things may compose the wrong way. This is not restrictive because we'll need it only for the Hat assembler (at the moment)

We combine them in a single function because there are avenues to optimize by recycling some computations (not completely exploited for now)

RigorousInvariantMeasures.refine_norms_of_powers — Method

Compute better and/or more estimates of power norms using the fact that ||Q^{k+h}|| ≤ ||Q^k|| * ||Q^h||. This uses multiplicativity, so it will not work for mixed norms, e.g., ||Q^k||{s → w}, or ||M^k|{U^0}|| (unless M preserves U^0, which is the case for Q|_{U^0}).

RigorousInvariantMeasures.relative_measure — Method

Relative measure of the intersection of (a,b) wrt the whole interval (c,d) Assumes that a,b and c,d are sorted correctly

RigorousInvariantMeasures.residualbound — Method

Return an upper bound to Q_h*w - w in the given norm

RigorousInvariantMeasures.residualboundnoise — Method

Return an upper bound to Q_h*w - w in the given norm

RigorousInvariantMeasures.rigorous_norm — Method

rigorous_norm(M; k = 10)

Computes the L² and L∞ norm of the powers of M|_V₀ up to the power k. The algorithm first computes a numeric estimate of the top Singular Value and then certifies it by an Interval Newton step

Outputs: v2

v₂ : it is the vector with the rigorous bounds on the L² norm these are obtained by observing that ||f||₂ = ||w||₂, where w is the vector of the Fourier coefficients of the trigonometric polynomial f

RigorousInvariantMeasures.two_grid_estimate — Method

Compute a two-grid error estimate.

The first return argument is the error, the second is the time breakdown according to ["dfly", "coarse", "assembling", "eigen", "norms", "error"]. (The sum of that vector is the total time taken)

RigorousInvariantMeasures.z_times_conjz — Method

Computes a rigorous upper bound for z*z'

Index

RigorousInvariantMeasures.CoarseGridQuantities

RigorousInvariantMeasures.ComposedDynamic

RigorousInvariantMeasures.Dual

RigorousInvariantMeasures.FineGridQuantities

RigorousInvariantMeasures.Fourier1D

RigorousInvariantMeasures.Hat

RigorousInvariantMeasures.Hat

RigorousInvariantMeasures.HatFunction

RigorousInvariantMeasures.HatFunctionOnTorus

RigorousInvariantMeasures.IntervalOnTorus

RigorousInvariantMeasures.NonIntegralPreservingDiscretizedOperator

RigorousInvariantMeasures.NormCacher

RigorousInvariantMeasures.NormCacher

RigorousInvariantMeasures.Ulam

RigorousInvariantMeasures._eval_T

RigorousInvariantMeasures.abs_or_mag

RigorousInvariantMeasures.add_column!

RigorousInvariantMeasures.assemble

RigorousInvariantMeasures.boundnorm

RigorousInvariantMeasures.compute_coarse_grid_quantities

RigorousInvariantMeasures.compute_fine_grid_quantities

RigorousInvariantMeasures.dfly

RigorousInvariantMeasures.distance_from_invariant

RigorousInvariantMeasures.distance_from_invariant_noise

RigorousInvariantMeasures.eig_costants_small_matrix

RigorousInvariantMeasures.eval_Clenshaw_BackwardFirst

RigorousInvariantMeasures.faster_certify

RigorousInvariantMeasures.finepowernormbounds

RigorousInvariantMeasures.finepowernormboundsnoise

RigorousInvariantMeasures.first_overlapping

RigorousInvariantMeasures.gamma

RigorousInvariantMeasures.get_norm

RigorousInvariantMeasures.get_norm

RigorousInvariantMeasures.infinite_sum_norms

RigorousInvariantMeasures.invariant_vector

RigorousInvariantMeasures.invariant_vector_noise

RigorousInvariantMeasures.last_overlapping

RigorousInvariantMeasures.max_nonzeros_per_row

RigorousInvariantMeasures.norm_2_estimator_square_nonrig

RigorousInvariantMeasures.norms_of_powers

RigorousInvariantMeasures.norms_of_powers_abstract_noise

RigorousInvariantMeasures.norms_of_powers_dfly

RigorousInvariantMeasures.norms_of_powers_from_coarser_grid

RigorousInvariantMeasures.norms_of_powers_from_coarser_grid_noise

RigorousInvariantMeasures.norms_of_powers_from_coarser_grid_noise

RigorousInvariantMeasures.norms_of_powers_noise

RigorousInvariantMeasures.norms_of_powers_trivial

RigorousInvariantMeasures.norms_of_powers_trivial_noise

RigorousInvariantMeasures.norms_of_powers_trivial_noise

RigorousInvariantMeasures.one_grid_estimate

RigorousInvariantMeasures.powernormbounds

RigorousInvariantMeasures.powernormbounds

RigorousInvariantMeasures.preimages

RigorousInvariantMeasures.preimages_and_derivatives

RigorousInvariantMeasures.refine_norms_of_powers

RigorousInvariantMeasures.relative_measure

RigorousInvariantMeasures.residualbound

RigorousInvariantMeasures.residualboundnoise

RigorousInvariantMeasures.rigorous_norm

RigorousInvariantMeasures.two_grid_estimate

RigorousInvariantMeasures.z_times_conjz