Basis

These are some of then basis implemented in the package.

Ulam Basis

RigorousInvariantMeasures.Ulam — Type

Equispaced Ulam basis on [0,1] of size n

Hat basis

RigorousInvariantMeasures.Hat — Type

This type represents a Hat basis on $S^1$. It contains a vector with the midpoints of the hats.

RigorousInvariantMeasures.Hat — Method

This constructs a Hat basis on $S^1$ on equispaced points

Base.length — Method

Return the size of the Hat basis

Return the size of the Hat basis

Hat functions types

Missing docstring.

Missing docstring for HatFunctionOnTorus. Check Documenter's build log for details.