We present some of the basis already implemented in the package

The Ulam basis

The hat basis on $S^1$

RigorousInvariantMeasures.Hat — Type

This type represents a Hat basis on $S^1$. It contains a vector with the midpoints of the hats.

RigorousInvariantMeasures.Hat — Method

This constructs a Hat basis on $S^1$ on equispaced points

RigorousInvariantMeasures.HatFunction — Type

Hat function (on the reals)

This is a piecewise linear function such that: f(x) = 0 if x <= lo f(mi) = 1 f(hi)

RigorousInvariantMeasures.HatFunctionOnTorus — Type

Hat function (on the torus)

This is a piecewise linear function such that: f(x) = 0 if x <= lo f(mi) = 1 f(x) = 0 if x >= ho

RigorousInvariantMeasures.IntervalOnTorus — Type

A separate type for intervals on the torus (mod 1) to "remind" us of the quotient

The interval is normalized in the constructor: the caller may assume that

0 <= i.lo < 1

i.hi < i.lo + 1 OR i==Interval(0,1)

Base.getindex — Method

Base.getindex(B::Hat, i::Int)

Make so that B[j] returns a HatFunctionOnTorus with the j-th basis element

Base.getindex — Method

makes so that B[j] returns a HatFunction with the j-th basis element

Base.iterate — Method

Given a preimage y of a point x, this iterator returns \phi_j(y)/T'(y)

Base.iterate — Method

Given a preimage y of a point x, this iterator returns \phi_j(y)/T'(y)

Base.length — Method

Return the size of the Hat basis

Base.length — Method

Return the size of the Hat basis

RecipesBase.apply_recipe — Method

Plots a function in the Hat basis

RecipesBase.apply_recipe — Method

Displays error on a function in the Hat basis

RecipesBase.apply_recipe — Method

Plots a function in the Hat basis

RecipesBase.apply_recipe — Method

Displays error on a function in the Hat basis

RigorousInvariantMeasures.BasisDefinition.nonzero_on — Method

BasisDefinition.nonzero_on(B::Hat, dual_element)

Return the range of indices of the elements of the basis whose support intersects with the given dual element (i.e., a pair (y, absT')). The range may end with length(B)+1; this must be interpreted "mod length(B)": it means that it intersects with the hat function peaked in 0 as well (think for instance y = 0.9999).

RigorousInvariantMeasures.BasisDefinition.nonzero_on — Method

Return the range of indices of the elements of the basis whose support intersects with the given dual element (i.e., a pair (y, absT')). The range may end with length(B)+1; this must be interpreted "mod length(B)": it means that it intersects with the hat function peaked in 0 as well (think for instance y = 0.9999).

The hat basis on $[0,1]$

The Chebyshev basis

Base.getindex — Method

makes so that B[j] returns a HatFunctionOnTorus with the j-th basis element

Base.length — Method

Return the size of the Chebyshev basis

RigorousInvariantMeasures._eval_T — Method

Eval the Chebyshev polynomial up to degree n on an array of points in [-1, 1].

Not satisfactory, the intervals explode

RigorousInvariantMeasures.eval_Clenshaw_BackwardFirst — Method

Eval a polynomial in Chebyshev basis, ClenshawBackward, using ball arithmetic Following Viviane Ledoux, Guillaume Moroz "Evaluation of Chebyshev polynomials on intervals andapplication to root finding"

RigorousInvariantMeasures.get_norm — Method

Return the norm of the matrix the NormCacher is working on.