We present some of the basis already implemented in the package

Abstract basis

RigorousInvariantMeasures.Dual — Type

Replacement of DualComposedWithDynamic.

RigorousInvariantMeasures.BasisDefinition.AverageZero — Type

AverageZero{B<:Basis}

Yield a basis of the space of average zero vectors

RigorousInvariantMeasures.BasisDefinition.aux_normalized_projection_error — Method

aux_normalized_projection_error(B::Basis)

Return a constant Eh (typically scales as h ~ 1/n) such that |||P_h f|||\leq |||f|||+ Eh * ||f||_s Must be rounded up correctly!

RigorousInvariantMeasures.BasisDefinition.aux_weak_bound — Method

aux_weak_bound(B::Basis)

Return a constant $M₂$ such that for a vector $v ∈ Uₕ$ |||v|||\leq M_2||v|| Must be rounded up correctly!

RigorousInvariantMeasures.BasisDefinition.bound_linalg_norm_L1_from_weak — Method

bound_linalg_norm_L1_from_weak(B::Basis)

Return a constant $A$ such that for a vector $v ∈ Uₕ$ ||v||_1\leq A||v|| Must be rounded up correctly!

RigorousInvariantMeasures.BasisDefinition.bound_linalg_norm_L∞_from_weak — Method

bound_linalg_norm_L∞_from_weak(B::Basis)

Return a constant $A$ such that for a vector $v ∈ Uₕ$ ||v||_\infty \leq A||v|| Must be rounded up correctly!

RigorousInvariantMeasures.BasisDefinition.bound_weak_norm_abstract — Function

bound_weak_norm_abstract(B::Basis, D=nothing; dfly_coefficients=nothing)

Returns an a priori bound on the weak norm of the abstract operator $L$

RigorousInvariantMeasures.BasisDefinition.bound_weak_norm_from_linalg_norm — Method

bound_weak_norm_from_linalg_norm(B::Basis)

Return constants W₁, W₂ such that for a vector $v ∈ Uₕ$ ||v||\leq W_1||v||_1+W_2||v||_{\infty} Must be rounded up correctly!

RigorousInvariantMeasures.BasisDefinition.evaluate — Method

evaluate(B::Basis, i, x)

Evaluate the i-th basis element at x

RigorousInvariantMeasures.BasisDefinition.evaluate_integral — Method

evaluate_integral(B::Basis, i; T = Float64)

Value of the integral on [0,1] of the i-th basis element

RigorousInvariantMeasures.BasisDefinition.integral — Method

integral(B::Basis, v; T = Float64)

Return the integral of the function with coefficients v in the basis B

RigorousInvariantMeasures.BasisDefinition.integral_covector — Method

integral_covector(B::Basis)

Return a covector that represents the integral in the basis B

RigorousInvariantMeasures.BasisDefinition.invariant_measure_strong_norm_bound — Method

invariant_measure_strong_norm_bound(B::Basis, D::Dynamic)

Bounds $||u||_s$, where $u$ is the invariant measure normalized with $i(u)=1$.

RigorousInvariantMeasures.BasisDefinition.is_integral_preserving — Method

is_integral_preserving(B::Basis)

Integral-preserving discretizations may specialize this to "true"

RigorousInvariantMeasures.BasisDefinition.is_refinement — Method

is_refinement(Bfine::Basis, Bcoarse::Basis)

Check if Bfine is a refinement of Bcoarse

Example

julia> using RigorousInvariantMeasures

julia> B = Ulam(1024)
Ulam{LinRange{Float64, Int64}}(LinRange{Float64}(0.0, 1.0, 1025))

julia> Bfine = Ulam(2048)
Ulam{LinRange{Float64, Int64}}(LinRange{Float64}(0.0, 1.0, 2049))

julia> is_refinement(Bfine, B)
true

julia> Bfine = Ulam(2049)
Ulam{LinRange{Float64, Int64}}(LinRange{Float64}(0.0, 1.0, 2050))

julia> is_refinement(Bfine, B)
false

RigorousInvariantMeasures.BasisDefinition.one_vector — Method

one_vector(B::Basis)

Vector that represents the function 1 in the basis B

RigorousInvariantMeasures.BasisDefinition.strong_norm — Method

strong_norm(B::Basis)

Return the type of the strong norm of the basis

Example

julia> using RigorousInvariantMeasures

julia> B = Ulam(1024)
Ulam{LinRange{Float64, Int64}}(LinRange{Float64}(0.0, 1.0, 1025))

julia> strong_norm(B)
TotalVariation

RigorousInvariantMeasures.BasisDefinition.strong_weak_bound — Method

strong_weak_bound(B::Basis)

Return a constant $M₁n$ such that for a vector $v ∈ Uₕ$ ||v||_s\leq M1n*||v|| Must be rounded up correctly!

RigorousInvariantMeasures.BasisDefinition.weak_by_strong_and_aux_bound — Method

Return constants ``S₁, S₂`` such that for a vector ``v ∈ Uₕ`` 
```||v||\leq S_1||v||_s+S_2|||v|||```

Must be rounded up correctly!

RigorousInvariantMeasures.BasisDefinition.weak_norm — Method

weak_norm(B::Basis)

Return the type of the weak norm of the basis

Example

julia> using RigorousInvariantMeasures

julia> B = Ulam(1024)
Ulam{LinRange{Float64, Int64}}(LinRange{Float64}(0.0, 1.0, 1025))

julia> weak_norm(B)
L1

RigorousInvariantMeasures.BasisDefinition.weak_projection_error — Method

weak_projection_error(B::Basis)

Return a constant Kh (typically scales as h ~ 1/n) such that ||P_h f-f||\leq Kh ||f||_s Must be rounded up correctly! This function is not exported explictly but is used in all the estimates.

Example

```jldoctest julia> using RigorousInvariantMeasures;

julia> B = Ulam(1024) Ulam{LinRange{Float64, Int64}}(range(0.0, stop=1.0, length=1025))

julia> RigorousInvariantMeasures.BasisDefinition.weakprojectionerror(B)

0.00048828125 ````

The Ulam basis

RigorousInvariantMeasures.Ulam — Type

Ulam

Ulam basis on [0,1] associated to the partition $p = \{x_0 = 0, x_1, \ldots, x_n=1\}$

RigorousInvariantMeasures.Ulam — Method

Ulam(n::Integer)

Equispaced Ulam basis on [0,1] of size n

Base.getindex — Method

Base.getindex(B::Ulam, i::Int)

Returns the i-th element of the Ulam basis as a function.

Example

julia> using RigorousInvariantMeasures

julia> B = Ulam(16)
Ulam{LinRange{Float64, Int64}}(LinRange{Float64}(0.0, 1.0, 17))

julia> B[1](1/32)
1

julia> B[2](1/32)
0

Base.iterate — Method

iterate(S::AverageZero{Ulam{T}}, state = 1) where{T}

Return the elements of the basis of average $0$ functions in the Ulam Basis

Base.iterate — Method

Given a preimage of an interval I_i, this iterator returns its relative intersection with all the elements of the Ulam basis that have nonzero intersection with it

Base.length — Method

Base.length(B::Ulam)

Returns the size of the Ulam basis (the size of the underlying vector -1)

Base.length — Method

Base.length(S::AverageZero{Ulam})

Return the size of the Average Zero space

RecipesBase.apply_recipe — Function

Displays error on a function in the Ulam basis

The w argument is unused, but kept for compatibility with other functions for different bases

RecipesBase.apply_recipe — Method

Plots a function in the Ulam basis

RigorousInvariantMeasures.BasisDefinition.nonzero_on — Method

nonzero_on(B::Ulam, (a, b))

Returns the indices of the elements of the Ulam basis that intersect with the interval y We do not assume an order of a and b; this should not matter unless the preimages are computed with very low accuracy. We assume, though, that y comes from the (possibly inexact) numerical approximation of an interval in $[0,1]$, i.e., we restrict to $y \cap [0,1]$

RigorousInvariantMeasures.relative_measure — Method

relative_measure((a,b)::Tuple{<:Interval,<:Interval}, (c,d)::Tuple{<:Interval,<:Interval})

Relative measure of the intersection of (a,b) wrt the whole interval (c,d) Assumes that a,b and c,d are sorted correctly

The hat basis on $S^1$

RigorousInvariantMeasures.Hat — Type

Hat

This type represents a Hat basis on $S^1$. It contains a vector with the midpoints of the hats.

RigorousInvariantMeasures.Hat — Method

Hat(n::Integer)

This constructs a Hat basis on $S^1$ on equispaced points

RigorousInvariantMeasures.HatFunction — Type

Hat function (on the reals)

This is a piecewise linear function such that: f(x) = 0 if x <= lo f(mi) = 1 f(hi)

RigorousInvariantMeasures.HatFunction — Method

Evaluate a HatFunction (on the real line).

Must work correctly when S is an interval.

RigorousInvariantMeasures.HatFunctionOnTorus — Type

HatFunctionOnTorus

Hat function (on the torus)

This is a piecewise linear function such that: f(x) = 0 if x <= lo f(mi) = 1 f(x) = 0 if x >= ho

RigorousInvariantMeasures.HatFunctionOnTorus — Method

Evaluate a HatFunctionOnTorus correctly on an IntervalOnTorus

Assumption: this is only called on functions defined on our partition, so either mi==0, hi==0, or the three values are in increasing order

RigorousInvariantMeasures.IntervalOnTorus — Type

IntervalOnTorus

A separate type for intervals on the torus (mod 1) to "remind" us of the quotient

The interval is normalized in the constructor: the caller may assume that

0 <= i.lo < 1

i.hi < i.lo + 1 OR i==Interval(0,1)

Base.getindex — Method

Base.getindex(B::Hat, i::Int)

Make so that B[j] returns a HatFunctionOnTorus with the j-th basis element

Base.getindex — Method

Base.getindex(B::HatNP, i::Int)

makes so that B[j] returns a HatFunction with the j-th basis element

Base.iterate — Method

Given a preimage y of a point x, this iterator returns \phi_j(y)/T'(y)

Base.iterate — Method

Given a preimage y of a point x, this iterator returns \phi_j(y)/T'(y)

Base.length — Method

Base.length(B::HatNP)

Return the size of the Hat basis

Base.length — Method

Base.length(B::Hat{T})

Return the size of the Hat basis

RecipesBase.apply_recipe — Method

Plots a function in the Hat basis

RecipesBase.apply_recipe — Method

Displays error on a function in the Hat basis

RecipesBase.apply_recipe — Method

Plots a function in the Hat basis

RecipesBase.apply_recipe — Method

Displays error on a function in the Hat basis

RigorousInvariantMeasures.BasisDefinition.nonzero_on — Method

BasisDefinition.nonzero_on(B::Hat, dual_element)

Return the range of indices of the elements of the basis whose support intersects with the given dual element (i.e., a pair (y, absT')). The range may end with length(B)+1; this must be interpreted "mod length(B)": it means that it intersects with the hat function peaked in 0 as well (think for instance y = 0.9999).

RigorousInvariantMeasures.BasisDefinition.nonzero_on — Method

Return the range of indices of the elements of the basis whose support intersects with the given dual element (i.e., a pair (y, absT')). The range may end with length(B)+1; this must be interpreted "mod length(B)": it means that it intersects with the hat function peaked in 0 as well (think for instance y = 0.9999).

The hat basis on $[0,1]$

The C2 basis

RigorousInvariantMeasures.C2BasisDefinition.EquispacedPartitionInterval — Type

Equispaced partition of size n of [0,1]

Base.iterate — Method

Return (in an iterator) the pairs (i, (x, |T'(x)|)) where x is a preimage of p[i], which describe the "dual" L* evaluation(p[i])

Base.iterate — Method

Given a preimage y of a point x, this iterator returns \phi_j(y)/T'(y)

Base.length — Method

Return the size of the C2 basisBase.length(S::AverageZero) = length(S.basis)-1

RigorousInvariantMeasures.BasisDefinition.nonzero_on — Method

Return the range of indices of the elements of the basis whose support intersects with the given dual element (i.e., a pair (y, absT', derder)).

The Chebyshev basis

Base.getindex — Method

Base.getindex(B::Chebyshev, i::Int)

Make so that B[j] returns a HatFunctionOnTorus with the j-th basis element

Base.length — Method

Return the size of the Chebyshev basis

RigorousInvariantMeasures._eval_T — Method

_eval_T

Eval the Chebyshev polynomial up to degree n on an array of points in [-1, 1].

Not satisfactory, the intervals explode

RigorousInvariantMeasures.eval_Clenshaw_BackwardFirst — Method

eval_Clenshaw_BackwardFirst

Eval a polynomial in Chebyshev basis, ClenshawBackward, using ball arithmetic Following Viviane Ledoux, Guillaume Moroz "Evaluation of Chebyshev polynomials on intervals andapplication to root finding"

RigorousInvariantMeasures.get_norm — Method

Return the norm of the matrix the NormCacher is working on.